Refactor: change file name (CurrentUser -> UserCurrent) & move to user (UserPatchReq: request -> user)
